backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Vector search
faiss-cpu==1.7.4
# pinecone-client==3.0.0  # Optional: uncomment for Pinecone

# AI & LangChain
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.20
openai==1.10.0
tiktoken==0.5.2

# Free AI Alternatives
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.37.2
huggingface-hub==0.20.3
openai-whisper==20231117

# Transcription APIs
deepgram-sdk==3.2.7

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'

# Audio/Video processing
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Redis (optional)
redis==5.0.1
hiredis==2.3.2

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.26.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# Code quality
ruff==0.1.14
black==24.1.1
mypy==1.8.0

# Monitoring (optional)
prometheus-client==0.19.0
